+++ Moco/Moco/MocoCasADiSolver/CasOCFunction.cpp
@@ -179,12 +179,20 @@ casadi::Sparsity Endpoint::get_sparsity_in(casadi_int i) {
         return casadi::Sparsity(0, 0);
     }
 }
-VectorDM Endpoint::eval(const VectorDM& args) const {
+VectorDM Cost::eval(const VectorDM& args) const {
     Problem::CostInput input{args.at(0).scalar(), args.at(1), args.at(2),
             args.at(3), args.at(4), args.at(5).scalar(), args.at(6), args.at(7),
             args.at(8), args.at(9), args.at(10), args.at(11).scalar()};
     VectorDM out{casadi::DM(casadi::Sparsity::dense(m_numEquations, 1))};
     m_casProblem->calcCost(m_index, input, out.at(0));
+    return out;
+}
+VectorDM EndpointConstraint::eval(const VectorDM& args) const {
+    Problem::CostInput input{args.at(0).scalar(), args.at(1), args.at(2),
+                             args.at(3), args.at(4), args.at(5).scalar(), args.at(6), args.at(7),
+                             args.at(8), args.at(9), args.at(10), args.at(11).scalar()};
+    VectorDM out{casadi::DM(casadi::Sparsity::dense(m_numEquations, 1))};
+    m_casProblem->calcEndpointConstraint(m_index, input, out.at(0));
     return out;
 }